+++ src/index.js
@@ -1,7 +1,8 @@
 import './sass/main.scss';
 import { axiosFirstFetchFn, axiosSecondFetchFn } from './js/filmApi';
-import { makeFilmsBox, fetchGenres } from './js/main';
+import { makeFilmsBox } from './js/main';
 axiosFirstFetchFn();
 axiosSecondFetchFn();
+makeFilmsBox();
 
 console.log(123);
