+++ src/js/filmApi.js
@@ -1,5 +1,7 @@
 import axios from 'axios';
 
+const API_KEY = '95f474a01cc4252905d63c7d958d5749';
+
 // FIRST METHOD
 axios.defaults.baseURL = 'https://api.themoviedb.org/3/movie/popular';
 axios.defaults.params = {
@@ -33,8 +35,24 @@ export const axiosSecondFetchFn = async () => {
   try {
     const response = await axios.request(options);
     const data = await response.data;
-    return console.log(data);
+    return data;
   } catch (error) {
     console.log(error);
+  }
+};
+
+//FETCHING ALL GENRES LIST
+
+export const fetchGenres = async () => {
+  try {
+    const response = await axios.get('https://api.themoviedb.org/3/genre/movie/list', {
+      params: {
+        api_key: API_KEY,
+      },
+    });
+    const genres = response.data.genres;
+    return genres;
+  } catch (error) {
+    console.error(error.message);
   }
 };
